id til int

+++ src/main/java/demo/BestClass.java
@@ -2,5 +2,5 @@ package demo;
 
 public class BestClass {
 
-    private double id;
+    private int id;
 }
